Added Command Group for Shooter Auto-Aiming and Firing
Command group, which would call all the necessary commands for
auto-aiming has been created. This includes rotating to face the boiler,
calculating the necessary shooter speed given a distance from the
boiler, and turning on the feeder.

In order to support the future calculation of the optimal feed speed
given the shooter speed, the feed command has been modified to accept a
direct speed input while still maintaining its ability to toggle on/off
for normal operation.